% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_sketcher.R
\name{chemdoodle_sketcherOutput}
\alias{chemdoodle_sketcherOutput}
\title{Widget output function for use in Shiny}
\usage{
chemdoodle_sketcherOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}
