% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_slideshow.R
\name{chemdoodle_slideshowOutput}
\alias{chemdoodle_slideshowOutput}
\title{Widget output function for use in Shiny}
\usage{
chemdoodle_slideshowOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}
