% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_transform.R
\name{chemdoodle_transform}
\alias{chemdoodle_transform}
\title{ChemDoodle Transform Canvas}
\usage{
chemdoodle_transform(smiles, width = 500, height = 500, bondscale = 14.4,
  backgroundColor = "white", scale = 1, atoms_font_size_2D = 12,
  atoms_font_families_2D = c("Helvetica", "Arial", "Dialog"),
  atoms_font_italic_2D = FALSE, atoms_circleDiameter_2D = 10,
  atoms_useJMOLColors = FALSE, atoms_usePYMOLColors = FALSE,
  atoms_implicitHydrogens_2D = TRUE,
  atoms_displayTerminalCarbonLabels_2D = TRUE,
  atoms_displayAllCarbonLabels_2D = FALSE, bonds_display = TRUE,
  bonds_width_2D = 1, bonds_saturationWidth_2D = 0.2,
  bonds_ends_2D = "round", bonds_useJMOLColors = FALSE,
  bonds_usePYMOLColors = FALSE, bonds_wedgeThickness_2D = 0.22,
  bonds_hashWidth_2D = 1)
}
\description{
Widget that accepts a SMILES string and visualizes it using ChemDoodle
}
