% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_transform.R
\name{chemdoodle_transformOutput}
\alias{chemdoodle_transformOutput}
\title{Widget output function for use in Shiny}
\usage{
chemdoodle_transformOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}
