% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_viewer.R
\name{chemdoodle_viewerOutput}
\alias{chemdoodle_viewerOutput}
\title{Widget output function for use in Shiny}
\usage{
chemdoodle_viewerOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}
