% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_json_utilities.R
\name{createAtom}
\alias{createAtom}
\title{make a CDK atom from ChemDoodle JSON}
\usage{
createAtom(atom)
}
\description{
a (Array) : Array of atom objects.
  [
  {
    i (String) : A unique identifier for this atom, only needs to be provided if necessary
    ,l (String, default='C') : Label
    ,x (Number, required) : X coordinate
    ,y (Number, required) : Y coordinate
    ,z (Number, default=0) : Z coordinate
    ,c (Integer, default=0) : Charge
    ,m (Integer, default=-1) : Mass
    ,r (Integer, default=0) : Number of radicals
    ,p (Integer, default=0) : Number of lone pairs
    ,q (Object, default=undefined) : If present, this object defines the query for the atom, as described by the query section below
  }, 
  ]
}
\keyword{internal}
