% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_utils.R
\name{mol_from_smiles}
\alias{mol_from_smiles}
\title{mol_from_smiles}
\usage{
mol_from_smiles(smiles)
}
\value{
IAtomContainer with coordinates
}
\description{
get an AtomContainer with 2d coords from a smiles string
}
\keyword{internal}

