% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_utils.R
\name{process_molecule}
\alias{process_molecule}
\title{process_molecule}
\usage{
process_molecule(mol)
}
\value{
list compatible with ChemDoodle Webcomponenets API
}
\description{
create JSON from an AtomContainer
}
\keyword{internal}
