% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_sketcher.R,
%   R/chemdoodle_slideshow.R, R/chemdoodle_transform.R, R/chemdoodle_viewer.R
\name{renderChemdoodle}
\alias{renderChemdoodle}
\alias{renderChemdoodle}
\alias{renderChemdoodle}
\alias{renderChemdoodle}
\title{Widget render function for use in Shiny}
\usage{
renderChemdoodle(expr, env = parent.frame(), quoted = FALSE)

renderChemdoodle(expr, env = parent.frame(), quoted = FALSE)

renderChemdoodle(expr, env = parent.frame(), quoted = FALSE)

renderChemdoodle(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny

Widget render function for use in Shiny

Widget render function for use in Shiny

Widget render function for use in Shiny
}
