% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdoodle_sketcher_gadgets.R
\name{drawMolecule}
\alias{drawMolecule}
\title{Initiates the ChemDoodle Sketcher}
\usage{
drawMolecule(mol = NULL, width = 600, height = 600)
}
\description{
helpful website:https://ryouready.wordpress.com/2013/11/20/sending-data-from-client-to-server-and-back-using-shiny/
}

